starting refactor for adding refactor hah
