+++ examples/vector_database.py
