+++ src/integrations/commenter_base.py
@@ -1,6 +1,6 @@
 from typing import List
 from abc import ABC, abstractmethod
-from code_review_tool import CodeComment
+from code_comment import CodeComment
 
 class CommenterBase(ABC):
     @abstractmethod
